working on auth module

+++ src/modules/daybook/views/EntryView.vue
@@ -7,7 +7,11 @@
                 <span class="mx-2 fs-4 fw-light">{{ yearDay }}</span>
             </div>
             <div>
-                <button v-if="entry.id" class="btn btn-danger mx-2" @click="onDeleteEntry">
+                <button
+                    v-if="entry.id"
+                    class="btn btn-danger mx-2"
+                    @click="onDeleteEntry"
+                >
                     Borrar
                     <i class="fa fa-trash-alt"></i>
                 </button>
@@ -37,6 +41,8 @@
 <script>
 import { defineAsyncComponent } from "vue";
 import { mapGetters, mapActions } from "vuex";
+import Swal from "sweetalert2";
+
 import getDayMonthYear from "../helpers/getDayMonthYear";
 export default {
     components: {
@@ -87,16 +93,40 @@ export default {
             this.entry = entry;
         },
         async saveEntry() {
+            new Swal({
+                title: "Espere por favor",
+                allowOutsideClick: false,
+            });
+            Swal.showLoading();
+
             if (this.entry.id) {
                 await this.updateEntry(this.entry);
             } else {
                 const id = await this.createEntry(this.entry);
                 this.$router.push({ name: "entry", params: { id } });
             }
+
+            Swal.fire("Guardado!", "Tu entrada se ha guardado", "success");
         },
-        onDeleteEntry() {
-            this.deleteEntry(this.entry.id);
-            this.$router.push({ name: "no-entry" });
+        async onDeleteEntry() {
+            const { isConfirmed } = await Swal.fire({
+                title: "¿Estás seguro?",
+                text: "¡No podrás revertir esto!",
+                showDenyButton: true,
+                confirmButtonText: "¡Sí, borrar!",
+            });
+            if (isConfirmed) {
+                new Swal({
+                    title: "Espere por favor",
+                    allowOutsideClick: false,
+                });
+                Swal.showLoading();
+
+                await this.deleteEntry(this.entry.id);
+                this.$router.push({ name: "no-entry" });
+
+                Swal.fire("Borrado!", "Tu entrada se ha borrado", "success");
+            }
         },
     },
     watch: {
